+++ cmake/Findudunits.cmake
@@ -13,6 +13,7 @@
 # The following paths will be searched in order if set in CMake (first priority) or environment (second priority):
 #
 #   - UDUNITS2_INCLUDE_DIRS & UDUNITS2_LIBRARIES - folders containing udunits2.h and libudunits2, respectively.
+#   - UDUNITS2_XML_PATH                          - path to udunits2.xml
 #   - UDUNITS2_ROOT                 - root of udunits installation
 #   - UDUNITS2_PATH                 - root of udunits installation
 #
@@ -28,6 +29,15 @@ find_path (
 		${UDUNITS2_PATH} $ENV{UDUNITS2_PATH}
   PATH_SUFFIXES include include/udunits2
 	DOC "Path to udunits2.h" )
+
+find_file (
+  udunits_XML_PATH
+  udunits2.xml
+  HINTS ${UDUNITS2_XML_PATH} $ENV{UDUNITS2_XML_PATH}
+    ${UDUNITS2_ROOT} $ENV{UDUNITS2_ROOT}
+    ${UDUNITS2_PATH} $ENV{UDUNITS2_PATH}
+  PATH_SUFFIXES share share/udunits
+  DOC "Path to udunits2.xml" )
 
 find_library(udunits_LIBRARY
 	NAMES udunits2 udunits
@@ -45,9 +55,9 @@ else()
 endif()
 
 include (FindPackageHandleStandardArgs)
-find_package_handle_standard_args (udunits DEFAULT_MSG udunits_LIBRARY udunits_INCLUDE_DIR)
+find_package_handle_standard_args (udunits DEFAULT_MSG udunits_LIBRARY udunits_INCLUDE_DIR udunits_XML_PATH)
 
-mark_as_advanced (udunits_LIBRARY udunits_INCLUDE_DIR)
+mark_as_advanced (udunits_LIBRARY udunits_INCLUDE_DIR udunits_XML_PATH)
 
 if(udunits_FOUND AND NOT TARGET udunits::udunits)
 	add_library(udunits::udunits INTERFACE IMPORTED)